Replace InvalidForeachVariableRule with VariableAssignNode-based rule
Create InvalidVariableAssignRule for VariableAssignNode that detects
$this reassignment in all contexts (regular assignments, foreach
variables, list destructuring). Extract the $this check from
InvalidAssignVarRule into this new unified rule.
