+++ src/Rules/Variables/InvalidVariableAssignRule.php
@@ -1,0 +1,44 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\Rules\Variables;
+
+use PhpParser\Node;
+use PHPStan\Analyser\Scope;
+use PHPStan\DependencyInjection\RegisteredRule;
+use PHPStan\Node\VariableAssignNode;
+use PHPStan\Rules\Rule;
+use PHPStan\Rules\RuleErrorBuilder;
+use function is_string;
+
+/**
+ * @implements Rule<VariableAssignNode>
+ */
+#[RegisteredRule(level: 0)]
+final class InvalidVariableAssignRule implements Rule
+{
+
+	public function getNodeType(): string
+	{
+		return VariableAssignNode::class;
+	}
+
+	public function processNode(Node $node, Scope $scope): array
+	{
+		$variable = $node->getVariable();
+		if (!is_string($variable->name)) {
+			return [];
+		}
+
+		if ($variable->name === 'this') {
+			return [
+				RuleErrorBuilder::message('Cannot re-assign $this.')
+					->identifier('assign.this')
+					->nonIgnorable()
+					->build(),
+			];
+		}
+
+		return [];
+	}
+
+}
