+++ tests/PHPStan/Rules/Variables/data/bug-14349.php
@@ -2,20 +2,42 @@
 
 namespace Bug14349;
 
-$a = [1];
+class Foo
+{
 
-foreach ($a as $this) {
-	var_dump($this);
+	/** @param array<int> $a */
+	public function doFoo(array $a): void
+	{
+		foreach ($a as $this) {
+			var_dump($this);
+		}
+
+		foreach ($a as &$this) {
+			var_dump($this);
+		}
+
+		foreach ($a as $this => $v) {
+			var_dump($this);
+		}
+
+		foreach ($a as $ok) {
+			var_dump($ok);
+		}
+
+		$this = 1;
+		$this = new self();
+		$this .= 'foo';
+		[$this] = [1];
+	}
+
+	public static function doBar(): void
+	{
+		$this = 1;
+	}
+
 }
 
-foreach ($a as &$this) {
-	var_dump($this);
-}
-
-foreach ($a as $this => $v) {
-	var_dump($this);
-}
-
-foreach ($a as $ok) {
-	var_dump($ok);
+function baz(): void
+{
+	$this = 1;
 }
